main/code/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.0.8
openai==1.30.1
tiktoken==0.6.0
pydantic==1.10.13
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0
dashscope==1.17.0

# memU Cloud API SDK (可选，用于集成 memU 云端服务)
# 注意：安装 memu-py 可能会更新一些依赖包（如 pydantic）
memu-py>=0.2.0